+++ concurrency-limits-servlet-jakarta/src/main/java/com/netflix/concurrency/limits/servlet/jakarta/ConcurrencyLimitServletFilter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.concurrency.limits.servlet;
+package com.netflix.concurrency.limits.servlet.jakarta;
 
 import com.netflix.concurrency.limits.Limiter;
