+++ concurrency-limits-servlet-jakarta/src/test/java/com/netflix/concurrency/limits/ConcurrencyLimitServletFilterSimulationTest.java
@@ -4,8 +4,8 @@ import com.netflix.concurrency.limits.executors.BlockingAdaptiveExecutor;
 import com.netflix.concurrency.limits.limit.FixedLimit;
 import com.netflix.concurrency.limits.limit.VegasLimit;
 import com.netflix.concurrency.limits.limiter.SimpleLimiter;
-import com.netflix.concurrency.limits.servlet.ConcurrencyLimitServletFilter;
-import com.netflix.concurrency.limits.servlet.ServletLimiterBuilder;
+import com.netflix.concurrency.limits.servlet.jakarta.ConcurrencyLimitServletFilter;
+import com.netflix.concurrency.limits.servlet.jakarta.ServletLimiterBuilder;
 import org.eclipse.jetty.servlet.FilterHolder;
 import org.junit.ClassRule;
 import org.junit.Ignore;
